# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library API",
    "version": "1.0.0",
    "description": "REST API for managing authors and books in MongoDB. All create and update operations validate their complete request bodies."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://cse-341-project2.onrender.com",
      "description": "Render deployment (update after creating the service)"
    }
  ],
  "tags": [
    {
      "name": "Authors",
      "description": "Author CRUD operations"
    },
    {
      "name": "Books",
      "description": "Book CRUD operations"
    },
    {
      "name": "System",
      "description": "Service status"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Check service health",
        "responses": {
          "200": {
            "description": "Service is available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get all authors",
        "responses": {
          "200": {
            "description": "Authors retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Create author",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "author created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResource"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A unique value already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get author by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "author retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "description": "Invalid MongoDB ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Authors"
        ],
        "summary": "Replace author",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "author updated successfully"
          },
          "400": {
            "description": "Invalid ID or request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A unique value already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "summary": "Delete author",
        "description": "Deletion is rejected while one or more books reference the author.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "author deleted successfully"
          },
          "400": {
            "description": "Invalid MongoDB ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "author not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Books still reference this author",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "Books retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "authorId",
            "in": "query",
            "required": false,
            "description": "Optionally filter books by author",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Create book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResource"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A unique value already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "book retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Invalid MongoDB ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Replace book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "book updated successfully"
          },
          "400": {
            "description": "Invalid ID or request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A unique value already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the resource",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$",
              "example": "64ac660864282d23d377e557"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "book deleted successfully"
          },
          "400": {
            "description": "Invalid MongoDB ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthorInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "firstName",
          "lastName",
          "birthDate",
          "nationality",
          "biography",
          "website"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Austen"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "1775-12-16"
          },
          "nationality": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "British"
          },
          "biography": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000,
            "example": "English novelist known for her social commentary."
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/authors/jane-austen"
          }
        }
      },
      "Author": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "firstName",
          "lastName",
          "birthDate",
          "nationality",
          "biography",
          "website",
          "_id",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Austen"
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "1775-12-16"
          },
          "nationality": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "British"
          },
          "biography": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000,
            "example": "English novelist known for her social commentary."
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/authors/jane-austen"
          },
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "64ac660864282d23d377e557"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "isbn",
          "genre",
          "publicationYear",
          "publisher",
          "language",
          "pageCount",
          "available",
          "description",
          "authorId"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200,
            "example": "Pride and Prejudice"
          },
          "isbn": {
            "type": "string",
            "example": "9780141439518"
          },
          "genre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Classic fiction"
          },
          "publicationYear": {
            "type": "integer",
            "minimum": 1000,
            "example": 1813
          },
          "publisher": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150,
            "example": "T. Egerton"
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "English"
          },
          "pageCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "example": 432
          },
          "available": {
            "type": "boolean",
            "example": true
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000,
            "example": "A novel about manners, upbringing, morality, and marriage."
          },
          "authorId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "64ac660864282d23d377e557"
          }
        }
      },
      "Book": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "isbn",
          "genre",
          "publicationYear",
          "publisher",
          "language",
          "pageCount",
          "available",
          "description",
          "authorId",
          "_id",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200,
            "example": "Pride and Prejudice"
          },
          "isbn": {
            "type": "string",
            "example": "9780141439518"
          },
          "genre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "example": "Classic fiction"
          },
          "publicationYear": {
            "type": "integer",
            "minimum": 1000,
            "example": 1813
          },
          "publisher": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150,
            "example": "T. Egerton"
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "English"
          },
          "pageCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "example": 432
          },
          "available": {
            "type": "boolean",
            "example": true
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000,
            "example": "A novel about manners, upbringing, morality, and marriage."
          },
          "authorId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "64ac660864282d23d377e557"
          },
          "_id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "64ac660864282d23d377e557"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatedResource": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$",
            "example": "64ac660864282d23d377e557"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "example": "title"
          },
          "message": {
            "type": "string",
            "example": "Is required and must be a string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      }
    }
  }
}